split Event* structs + impls into event.rs
