+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod display;
+pub mod event;
 pub mod widget;
 
 pub use euclid;
